+++ packages/changed-elements-react/src/widgets/comparisonJobWidget/components/VersionCompareSelectComponent.tsx
@@ -3,7 +3,7 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import { IModelConnection } from "@itwin/core-frontend";
-import { useState } from "react";
+import { ReactNode, useState } from "react";
 import { ProgressRadial } from "@itwin/itwinui-react";
 import { VersionCompareSelectorInner } from "./VersionCompareSelectorInner";
 import { CurrentNamedVersionAndNamedVersions } from "../models/NamedVersions";
@@ -24,7 +24,7 @@ export interface VersionCompareSelectorProps {
   /** Named Versions to be displayed */
   namedVersions: CurrentNamedVersionAndNamedVersions | undefined;
   /** Optional prop for a user supplied component to handle managing named versions.*/
-  manageNamedVersionsSlot?: () => React.ReactNode;
+  manageNamedVersionsSlot?: ReactNode | undefined;
 }
 
 /**